refactor(canvas): establish chat composer module
