+++ apps/canvas-workspace/src/renderer/src/agent-chat/composer/chatComposerDraftStore.ts
@@ -1,4 +1,4 @@
-import type { ChatImageAttachment } from '../types';
+import type { ChatImageAttachment } from '../../types';
 
 export interface ChatComposerDraft {
   input: string;
