+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatComposer/__tests__/useChatComposerInput.tabs.test.tsx
@@ -2,15 +2,15 @@
 import { act } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import { I18nProvider } from '../../../i18n';
-import type { AgentContextTabRef } from '../../../types';
-import { resetChatComposerDraftsForTests } from './chatComposerDraftStore';
-import { useMentions } from './useMentions';
-import { resetPluginMentionItemsForTests } from '../../../agent-chat/mentions/pluginMentionItems';
+import { I18nProvider } from '../../../../i18n';
+import type { AgentContextTabRef } from '../../../../types';
+import { resetChatComposerDraftsForTests } from '../../../../agent-chat/composer/chatComposerDraftStore';
+import { useChatComposerInput } from '../useChatComposerInput';
+import { resetPluginMentionItemsForTests } from '../../../../agent-chat/mentions/pluginMentionItems';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
 
-type Hook = ReturnType<typeof useMentions>;
+type Hook = ReturnType<typeof useChatComposerInput>;
 
 let root: Root | null = null;
 let host: HTMLDivElement | null = null;
@@ -28,7 +28,7 @@ const dockTab: AgentContextTabRef = {
 };
 
 const Probe = () => {
-  latest = useMentions({
+  latest = useChatComposerInput({
     agentScope: { kind: 'global' },
     allWorkspaces: [
       { id: 'workspace-product', name: 'Product' },
