+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatComposer/__tests__/useSubmitDomReviewComments.test.tsx
@@ -2,8 +2,8 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it, vi } from 'vitest';
-import type { AgentContextDomReviewComment, AgentRequestContext } from '../../../types';
-import { useSubmitDomReviewComments } from './useSubmitDomReviewComments';
+import type { AgentContextDomReviewComment, AgentRequestContext } from '../../../../types';
+import { useSubmitDomReviewComments } from '../useSubmitDomReviewComments';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
