+++ apps/canvas-workspace/src/renderer/src/components/chat/ChatComposer/submitQuickAction.ts
